+++ Tapjoy/CHANGELOG.md
@@ -1,4 +1,10 @@
 ## Changelog
+  * 12.4.2.0
+    * This version of the adapters has been certified with Tapjoy 12.4.2.
+
+  * 12.4.1.0
+    * This version of the adapters has been certified with Tapjoy 12.4.1.
+
   * 12.4.0.1
     * Log the Tapjoy placement ID in ad lifecycle events.